Separate the logic
